Big ol refactor Hopefully better file structure
