+++ duckbot/handlers/command/gamble.py
@@ -1,275 +1,12 @@
 # Python imports
-import re
-import shlex
-from datetime import datetime, date, time
+from datetime import datetime
 # Project imports
-import util
-from util import DiagMessage
-from util import Bank
-from gameHandlers import *
-
-# Handler instances
-global help_handler, roll_handler, gamble_handler
-
-# Help handler class
-class HelpHandler:
-#{{{
-
-    # Constructor for Help handler
-    # Params: bot_id - user id for the bot
-    # Return: HelpHandler instance
-    def __init__(self, bot_id):
-    #{{{
-        self.bot_id = bot_id
-        #{{{ - Help messages
-        self.help_messages = {
-             util.COMMANDS["HI"] :\
-                ("Legacy HI command\n"
-                "Usage: <@" + bot_id + "> HI")
-            ,util.COMMANDS["UPDATE"] :\
-                ("Causes the bot to shutdown and signal "
-                "the monitor script to check for updates\n"
-                "Usage: <@" + bot_id + "> UPDATE")
-            ,util.COMMANDS["HELP"] :\
-                "Don't get smart, you know how to use this"
-            ,util.COMMANDS["ROLL"] :\
-                ("Rolls dice based on parameters given\n"
-                "Usage: <@" + bot_id + "> ROLL ( [d]X | YdX )\n"
-                "Where X is the number of faces and Y is the number of dice")
-            ,util.COMMANDS["COIN"] :\
-                ("Flip a coin\n"
-                "Usage: <@" + bot_id + "> COIN")
-            ,util.COMMANDS["EIGHTBALL"] :\
-                ("Shake the magic 8ball\n"
-                "Usage: <@" + bot_id + "> 8BALL")
-            ,util.COMMANDS["FACTOID"] :\
-                ("Pull out a random and totally true fact\n"
-                "Usage: <@" + bot_id + "> FACTOID")
-            ,util.COMMANDS["PICKIT"] :\
-                ("Pick from a number of things\n"
-                "Usage: <@" + bot_id + "> PICKIT <item1> <item2> ...\n"
-                "Use quotes to have items with spaces in them :duck:")
-            ,util.COMMANDS["JOIN"] :\
-                ("Add yourself to the gambler's bank\n"
-                "Usage: <@" + bot_id + "> JOIN\n"
-                "Can only be used in gambling approved channels :duck:")
-            ,util.COMMANDS["CHECKBUX"] :\
-                ("Check bank balance of yourself or others\n"
-                "Usage: <@" + bot_id + "> CHECKBUX [target]\n"
-                "No target defaults to yourself :duck:")
-            ,util.COMMANDS["BET"] :\
-                ("Bet on a game with bank balance to win big\n"
-                "Usage: <@" + bot_id + "> BET <amount> <game> <game-options>\n"
-                "List of currently supported games: " + ", ".join(GAMES) + "\n"
-                "Use HELP BET <game> for details on options")
-        }
-        #}}}
-        #{{{ - Game help messages
-        self.game_help_messages = {
-             GAMES["COIN"] :\
-                ("Flip a coin and call it\n"
-                "Usage options: COIN ( H[EADS] | T[AILS] )")
-            ,GAMES["DICE"] :\
-                ("Roll the dice and guess even or odd\n"
-                "Usage options: DICE ( E[VENS] | O[DDS] )")
-        }
-        #}}}
-    #}}}
-
-    # Retrieve help message based on passed values
-    # Params: parms - list of strings to check for help commands
-    # Return: requested help messages or generic one if no specific
-    def act(self, parms):
-    #{{{
-        if parms:
-            command = util.COMMANDS.get(parms[0],0)
-            command = util.COMMANDS_ALT.get(parms[0],0) if not command else command
-            if command == util.COMMANDS["BET"] and len(parms) > 1:
-                subcommand = GAMES.get(parms[1],0)
-                response = self.game_help_messages.get(subcommand,
-                    parms[1] + " is not a recognized game")
-                return response
-            else:
-                response = self.help_messages.get(command,
-                    parms[0] + " is not a recognized command")
-                return response
-        else:
-            return ("Duckbot is a general purpose slackbot for doing various things\n"
-                    "To interact with it use <@" + self.bot_id + "> <command>\n"
-                    "Supported commands: " + ", ".join(util.COMMANDS) + "\n"
-                    "Use <@" + self.bot_id + "> HELP <command> for more details")
-    #}}}
-#}}}
-
-# Roll handler class
-class RollHandler:
-#{{{
-    ROLL_REGEX="^(:[a-z0-9_-]+:|\d+)?D(:[a-z0-9_-]+:|\d+)$"
-    #{{{ - CHARACTER_ROLLS
-    CHARACTER_ROLLS = [
-         ":DRAGON:"
-        ,"CHARACTER"
-        ,"CHAR"
-    ]
-    #}}}
-
-    # Constructor for Roll handler
-    # Params: None
-    # Return: RollHandler instance with ranges and regex initialized
-    def __init__(self):
-    #{{{
-        # Range from 1-100 allowed
-        self.die_range=range(1,101)
-        # Can pick from 2-20 things
-        self.pick_range=range(2,21)
-    #}}}
-
-    # Parse roll command parms and return values
-    # Params: roll_parms - list of roll parameters
-    # Return:  0, list of rolls
-    #          1, list of stats
-    #         -1, None for bad param
-    #         -2, None for no param
-    def roll(self, roll_parms):
-    #{{{
-        if not roll_parms:
-            return -2, None
-        # Check for character roll
-        elif roll_parms[0] in self.CHARACTER_ROLLS:
-            return 1, self.characterRoll()
-
-        results = re.search(self.ROLL_REGEX, roll_parms[0])
-
-        # If we matched a dX or YdX case
-        if results:
-        #{{{
-            # YdX case
-            if results.group(1):
-                # Numeric
-                try:
-                    die_num = int(results.group(1))
-                # Or emoji
-                except ValueError:
-                    die_num = util.EMOJI_ROLLS.get(results.group(1),-1)
-            else:
-                die_num = 1
-
-            # dX half
-            # Numeric
-            try:
-                die_size = int(results.group(2))
-            # Or emoji
-            except ValueError:
-                die_size = util.EMOJI_ROLLS.get(results.group(2),-1)
-        #}}}
-        # Just X case
-        else:
-        #{{{
-            die_num = 1
-            # Numeric
-            try:
-                die_size = int(roll_parms[0])
-            # Or emoji
-            except ValueError:
-                die_size = util.EMOJI_ROLLS.get(roll_parms[0],-1)
-        #}}}
-
-        # Check range for valid rolls
-        if (die_num in self.die_range and die_size in self.die_range):
-        #{{{
-            rolls = util.doRolls(die_size, die_num)
-            if die_num == 1:
-                rolls.append(self._emojiRating(rolls[0], die_size))
-            else:
-                rolls.append(sum(rolls))
-            return 0, rolls
-        #}}}
-        else:
-            return -1, None
-    #}}}
-
-    # Roll for dnd character stats
-    # Params: None
-    # Return: list of stat lists
-    def characterRoll(self):
-    #{{{
-        rolls = []
-        for i in range(0,6):
-            rolls.append(util.doRolls(6,4))
-        return rolls
-    #}}}
-
-    # Roll for coin flip
-    # Params: None
-    # Return: heads or tails
-    def coinRoll(self):
-    #{{{
-        result = util.doRolls(2)[0]
-        if result == 1:
-            return "HEADS"
-        else:
-            return "TAILS"
-    #}}}
-
-    # Roll for 8ball response
-    # Params: None
-    # Return: eight ball message
-    def eightballRoll(self):
-    #{{{
-        roll = util.doRolls(len(util.EIGHTBALL_RESPONSES))[0]
-        return util.EIGHTBALL_RESPONSES[roll]
-    #}}}
-
-    # Roll for pickit response
-    # Params: pick_parms - string to split and pick from
-    # Return: 0, chosen thing
-    #         1, range to pick from for error
-    def pickitRoll(self, pick_parms):
-    #{{{
-        try:
-            # Split on spaces while preserving quoted strings,
-            # then remove empty strings because people suck
-            pick_parms = shlex.split(" ".join(pick_parms))
-            pick_parms = [val for val in pick_parms if val != ""]
-        except ValueError:
-            return 2, None
-        if len(pick_parms) in self.pick_range:
-            return 0, pick_parms[util.doRolls(len(pick_parms))[0]-1]
-        else:
-            return 1, self.pick_range
-    #}}}
-
-    # Roll for factoid response
-    # Params: None
-    # Return: Generated factoid (nothing for right now)
-    def factoidRoll(self):
-        return "Not yet. Kweh :duck:"
-
-    # Give emoji ratings based on roll score
-    # Params: roll - roll score to rate
-    #         die  - total to compare score to
-    # Return: emoji string for rating
-    def _emojiRating(self, roll, die):
-    #{{{
-        if roll == 1:
-            return ":hyperbleh:"
-        elif roll == die:
-            return ":partyparrot:"
-        elif roll == 69:
-            return ":eggplant:"
-        elif roll == 420:
-            return ":herb:"
-        elif roll <= die/2:
-            return ":bleh:"
-        else:
-            return ":ok_hand:"
-    #}}}
-
-#}}}
+from util.bank import Bank
+from util import util
+import handlers.games as Games
 
 # Gambling handler class
 class GambleHandler:
-#{{{
     BAD_CHANNEL_MSG = ("Sorry, this is not an approved "
         "channel for gambling content\n Please keep it to "
         "channels with the :slot_machine: label :duck:")
@@ -294,7 +31,7 @@ class GambleHandler:
     # Return: GambleHandler instance with approved channels added
     def __init__(self, channels):
     #{{{
-        self.GAMES = GAMES
+        self.GAMES = Games.GAMES
         self.logger = util.logger
         self.approved_channels = self._getApproved(channels)
         self.bank = Bank()
@@ -566,10 +303,9 @@ class GambleHandler:
             self.bank.nuke()
         # Lost one
         elif roll < 50:
-            self.bank.
+            self.bank.removeOne()
         for key in self.GACHA_RANGES:
             if roll in key:
                 return self.GACHA_RANGES[key]
         return None
     #}}}
-#}}}
