+++ duckbot/util/diagCodes.py
